+++ 3rdParty/s2/geometry/s1angle.cc
@@ -33,7 +33,7 @@ ostream& operator<<(ostream& os, S1Angle const& a) {
   double degrees = a.degrees();
   char buffer[13];
   int sz = snprintf(buffer, sizeof(buffer), "%.7f", degrees);
-  if (sz >= 0 && sz < sizeof(buffer)) {
+  if (sz >= 0 && (size_t)sz < sizeof(buffer)) {
     return os << buffer;
   } else {
     return os << degrees;
