+++ 3rdParty/s2/geometry/s2polygon.cc
@@ -95,7 +95,7 @@ void S2Polygon::Release(vector<S2Loop*>* loops) {
 }
 
 static void DeleteLoopsInVector(vector<S2Loop*>* loops) {
-  for (int i = 0; i < loops->size(); ++i) {
+  for (size_t i = 0; i < loops->size(); ++i) {
     delete loops->at(i);
   }
   loops->clear();
@@ -124,7 +124,7 @@ bool S2Polygon::IsValid(const vector<S2Loop*>& loops) {
   // If a loop contains an edge AB, then no other loop may contain AB or BA.
   if (loops.size() > 1) {
     unordered_map<S2PointPair, pair<int, int> > edges;
-    for (int i = 0; i < loops.size(); ++i) {
+    for (size_t i = 0; i < loops.size(); ++i) {
       S2Loop* lp = loops[i];
       for (int j = 0; j < lp->num_vertices(); ++j) {
         S2PointPair key = make_pair(lp->vertex(j), lp->vertex(j + 1));
@@ -143,12 +143,12 @@ bool S2Polygon::IsValid(const vector<S2Loop*>& loops) {
 
   // Verify that no loop covers more than half of the sphere, and that no
   // two loops cross.
-  for (int i = 0; i < loops.size(); ++i) {
+  for (size_t i = 0; i < loops.size(); ++i) {
     if (!loops[i]->IsNormalized()) {
       VLOG(2) << "Loop " << i << " encloses more than half the sphere";
       return false;
     }
-    for (int j = i + 1; j < loops.size(); ++j) {
+    for (size_t j = i + 1; j < loops.size(); ++j) {
       // This test not only checks for edge crossings, it also detects
       // cases where the two boundaries cross at a shared vertex.
       if (loops[i]->ContainsOrCrosses(loops[j]) < 0) {
@@ -177,7 +177,7 @@ bool S2Polygon::IsValid(bool check_loops, int max_adjacent) const {
 void S2Polygon::InsertLoop(S2Loop* new_loop, S2Loop* parent,
                            LoopMap* loop_map) {
   vector<S2Loop*>* children = &(*loop_map)[parent];
-  for (int i = 0; i < children->size(); ++i) {
+  for (size_t i = 0; i < children->size(); ++i) {
     S2Loop* child = (*children)[i];
     if (child->ContainsNested(new_loop)) {
       InsertLoop(new_loop, child, loop_map);
@@ -191,7 +191,7 @@ void S2Polygon::InsertLoop(S2Loop* new_loop, S2Loop* parent,
   // Some of the children of the parent loop may now be children of
   // the new loop.
   vector<S2Loop*>* new_children = &(*loop_map)[new_loop];
-  for (int i = 0; i < children->size();) {
+  for (size_t i = 0; i < children->size();) {
     S2Loop* child = (*children)[i];
     if (new_loop->ContainsNested(child)) {
       new_children->push_back(child);
@@ -209,7 +209,7 @@ void S2Polygon::InitLoop(S2Loop* loop, int depth, LoopMap* loop_map) {
     loops_.push_back(loop);
   }
   vector<S2Loop*> const& children = (*loop_map)[loop];
-  for (int i = 0; i < children.size(); ++i) {
+  for (size_t i = 0; i < children.size(); ++i) {
     InitLoop(children[i], depth + 1, loop_map);
   }
 }
@@ -220,7 +220,7 @@ bool S2Polygon::ContainsChild(S2Loop* a, S2Loop* b, LoopMap const& loop_map) {
 
   if (a == b) return true;
   vector<S2Loop*> const& children = loop_map.find(a)->second;
-  for (int i = 0; i < children.size(); ++i) {
+  for (size_t i = 0; i < children.size(); ++i) {
     if (ContainsChild(children[i], b, loop_map)) return true;
   }
   return false;
@@ -675,7 +675,7 @@ class S2LoopsAsVectorsIndex: public S2LoopSequenceIndex {
     DecodeIndex(index, &loop_index, &vertex_in_loop);
     vector<S2Point> const* loop = loops_[loop_index];
     *from = &loop->at(vertex_in_loop);
-    *to = &loop->at(vertex_in_loop == loop->size() - 1
+    *to = &loop->at((size_t)vertex_in_loop == loop->size() - 1
                       ? 0
                       : vertex_in_loop + 1);
   }
@@ -769,7 +769,7 @@ static void ClipBoundary(S2Polygon const* a, bool reverse_a,
       DCHECK_EQ((b->Contains(a1) ^ invert_b), inside);
       if (inside) intersections.push_back(make_pair(1, a1));
       sort(intersections.begin(), intersections.end());
-      for (int k = 0; k < intersections.size(); k += 2) {
+      for (size_t k = 0; k < intersections.size(); k += 2) {
         if (intersections[k] == intersections[k+1]) continue;
         builder->AddEdge(intersections[k].second, intersections[k+1].second);
       }
@@ -859,7 +859,7 @@ vector<S2Point>* SimplifyLoopAsPolyline(S2Loop const* loop, S1Angle tolerance) {
   // Add them all except the last: it is the same as the first.
   vector<S2Point>* simplified_line = new vector<S2Point>(indices.size() - 1);
   VLOG(4) << "Now simplified to: ";
-  for (int i = 0; i + 1 < indices.size(); ++i) {
+  for (size_t i = 0; i + 1 < indices.size(); ++i) {
     (*simplified_line)[i] = line.vertex(indices[i]);
     VLOG(4) << S2LatLng(line.vertex(indices[i]));
   }
@@ -884,7 +884,7 @@ void BreakEdgesAndAddToBuilder(S2LoopsAsVectorsIndex* edge_index,
     ClipEdge(*from, *to, edge_index, false, &intersections);
     intersections.push_back(make_pair(1, *to));
     sort(intersections.begin(), intersections.end());
-    for (int k = 0; k + 1 < intersections.size(); ++k) {
+    for (size_t k = 0; k + 1 < intersections.size(); ++k) {
       if (intersections[k] == intersections[k+1]) continue;
       builder->AddEdge(intersections[k].second, intersections[k+1].second);
     }
@@ -929,7 +929,7 @@ void S2Polygon::InitToSimplified(S2Polygon const* a, S1Angle tolerance) {
     }
   }
 
-  for (int i = 0; i < simplified_loops.size(); ++i) {
+  for (size_t i = 0; i < simplified_loops.size(); ++i) {
     delete simplified_loops[i];
   }
   simplified_loops.clear();
@@ -971,7 +971,7 @@ void S2Polygon::InternalClipPolyline(bool invert,
     sort(intersections.begin(), intersections.end());
     // At this point we have a sorted array of vertex pairs representing
     // the edge(s) obtained after clipping (a0,a1) against the polygon.
-    for (int k = 0; k < intersections.size(); k += 2) {
+    for (size_t k = 0; k < intersections.size(); k += 2) {
       if (intersections[k] == intersections[k+1]) continue;
       S2Point const& v0 = intersections[k].second;
       S2Point const& v1 = intersections[k+1].second;
@@ -1034,7 +1034,7 @@ S2Polygon* S2Polygon::DestructiveUnionSloppy(vector<S2Polygon*>* polygons,
   // to the queue until we have a single polygon to return.
   typedef multimap<int, S2Polygon*> QueueType;
   QueueType queue;  // Map from # of vertices to polygon.
-  for (int i = 0; i < polygons->size(); ++i)
+  for (size_t i = 0; i < polygons->size(); ++i)
     queue.insert(make_pair((*polygons)[i]->num_vertices(), (*polygons)[i]));
   polygons->clear();
